Several suggestions
I think 'character' looks strange enough for newcomers to warrant its own entry in the table.

Likewise, people should be careful when they want to e.g. align numbers by prepending a zero: it changes the base of the number from decimal to octal!

s-expressions were not introduced before, so it is hard to remember anything about them at the point where you elaborate on Lists.