Update lesson-25-intro-to-core-async.md
`alts!` takes a vector of channels as its first argument. The time we use `alts!` in the Detecting Key Chords example the syntax is right. This time we specify the channels as separate arguments. Which means the second channel is interpreted as an option to `alts!` and we get a very puzzling error message.